+++ app_filegallery/src/main/kotlin/dev/ragnarok/filegallery/util/InstancesCounter.kt
@@ -1,22 +1,22 @@
 package dev.ragnarok.filegallery.util
 
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 class InstancesCounter {
-    private val map: MutableMap<Class<*>, AtomicInteger>
-    fun incrementAndGet(c: Class<*>): Int {
+    private val map: MutableMap<Class<*>, AtomicLong>
+    fun incrementAndGet(c: Class<*>): Long {
         var counter = map[c]
         if (counter == null) {
-            counter = AtomicInteger()
+            counter = AtomicLong()
             map[c] = counter
         }
         return counter.incrementAndGet()
     }
 
-    fun fireExists(c: Class<*>, id: Int) {
+    fun fireExists(c: Class<*>, id: Long) {
         var counter = map[c]
         if (counter == null) {
-            counter = AtomicInteger()
+            counter = AtomicLong()
             map[c] = counter
         }
         if (counter.get() < id) {
